Move working audio to new file
