+++ CM7/Core/Inc/audiohandler.h
@@ -1,0 +1,25 @@
+/*
+ * audiohandler.h
+ *
+ *  Created on: Jul 9, 2025
+ *      Author: milov
+ */
+
+#ifndef INC_AUDIOHANDLER_H_
+#define INC_AUDIOHANDLER_H_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "stm32h7xx_hal.h"
+#include "stm32h747i_discovery_audio.h"
+#include "main.h"
+
+void MicrophoneStartProcess();
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* INC_AUDIOHANDLER_H_ */
